requirements.txt
streamlit
python-dotenv
groq 
httpx